solution for String-Constructor under Algorithms/Strings
